tests: fix instance logger
There was a problem in `TestInstanceLogger` test with the fact that
cleanup after the test recieves error `io: read/write on closed pipe`.

After the patch the order of pipes closing was updated by moving it to
the cleanup block. A check was also added to send a SIGKILL signal to
the process if it is alive, to avoid the `sending a signal to a killed
process` error.

Part of #931

+++ cli/running/process_controller.go
@@ -102,17 +102,21 @@ func (pc *processController) StopWithSignal(waitTimeout time.Duration, stopSigna
 	// Terminate the process at any cost.
 	select {
 	case <-time.After(waitTimeout):
-		// Send "SIGKILL" signal
-		if err := pc.Cmd.Process.Kill(); err != nil {
-			return fmt.Errorf("failed to send SIGKILL to instance: %s", err)
-		} else {
-			// Wait for the process to terminate.
-			<-waitDone
-			return nil
+		if pc.IsAlive() {
+			// Send "SIGKILL" signal if process is still alive.
+			if err := pc.Cmd.Process.Kill(); err != nil {
+				return fmt.Errorf("failed to send SIGKILL to instance: %s", err)
+			} else {
+				// Wait for the process to terminate.
+				<-waitDone
+				return nil
+			}
 		}
 	case err := <-waitDone:
 		return err
 	}
+
+	return nil
 }
 
 // GetPid returns process PID.
